requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
numpy
pydantic
pytest
unittest-xml-reporting
    